+++ mskpy/image/core.py
@@ -1,22 +1,33 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 
 """
-image.core --- General image operations.
-========================================
+image.core --- General image operations, generators.
+====================================================
 
 .. autosummary::
    :toctree: generated/
 
-   imshift
-   rarray
-   rebin
-   stack2grid
-   tarray
-   unwrap
-   xarray
-   yarray
+   imshift - Shift an image.
+   rarray - Create an array of distances to a point.
+   rebin - Grow or shrink an array by an integer amount.
+   stack2grid - Tile a stack of images.
+   tarray - Create an array of angles with the same vertex.
+   unwrap - Transform an array from a rectangular to azimuthal projection.
+   xarray - Create an array of distances along a line.
+   yarray - Create an array of distances along a line.
 
 """
+
+__all__ = [
+    'imshift',
+    'rarray',
+    'rebin',
+    'stack2grid',
+    'tarray',
+    'unwrap',
+    'xarray',
+    'yarray'
+]
 
 import numpy as np
 
@@ -290,7 +301,7 @@ def refine_center(func, im, yx, N, subsample, scale=0, **kwargs):
 
 def unwrap(im, yx, dtdr=0, scale=None, bins=100, range=None,
            dtype=float):
-    """Transform image from rectangular to cylindrical projection.
+    """Transform image from rectangular to azimuthal projection.
 
     Parameters
     ----------